GH-3: Workaround `counter.` prefix explicitly
Resolves spring-cloud-stream-app-starters/counter#3

With Java 8 Spring Boot auto-configure `BufferCounterService` which
has an inconsistency with the `DefaultCounterService`, where the last
one checks for the `counter.` prefix, but `BufferCounterService` just
for the `counter`.
This is a problem because the `spring.application.name` is populated as
`counter-sink`, therefore we don't have `counter.` in case of `BufferCounterService`.
Having that the SCDF can't get access to our counter because it reads them as
`spring.metrics.counter.*`

* While Spring Boot is fixing consistency with the `counter.` prefix, we provide
a workaround with an explicit setting `counter.` prefix in the `CounterSinkConfiguration`